impoved serial communication
