+++ spi_functions.cpp
@@ -2,7 +2,7 @@
 #include <spi_functions.h>
 
 ////////////////////////////////////////////////////////
-int sendConfiguration(Serial &serialOut,  SPI &dac_port, DigitalOut &sync)
+int sendConfiguration(SPI &dac_port, DigitalOut &sync, bool scan_mode)
 {
     int nWords= 12;
     int returnedData;
@@ -27,10 +27,12 @@ int sendConfiguration(Serial &serialOut,  SPI &dac_port, DigitalOut &sync)
     dac_data[11]= 0x0401;
     for (int i = 0; i < nWords; i++)
     {
-        serialOut.printf("WRITING %d (0x%04x).\n", dac_data[i], dac_data[i]);
+        if (!scan_mode)
+            printf("WRITING %d (0x%04x).\n", dac_data[i], dac_data[i]);
         returnedData = mySPISend(dac_data[i], dac_port, sync);
         //returnedData = mySPIRead(serialOut, dac_data[i], dac_port, sync);
-        serialOut.printf("Read back 0x%04x \n" , returnedData );
+        if (!scan_mode)
+            printf("Read back 0x%04x \n" , returnedData );
     }
     return 0;
 }
@@ -49,37 +51,37 @@ int mySPISend( unsigned short data, SPI &dac_port, DigitalOut &sync)
     unsigned short return1, return2;
     
     //sync = 1;
-    wait (0.001); // Changed from 0.5, DGC, 1 Oct 18
+    wait_us(1000); // Changed from 0.5, DGC, 1 Oct 18
 
     sync = 0; // drive chip select
     
-    wait(0.0001); // wait 100us for select line to go fully low after selecting chip . 
+    wait_us(1000); // wait 100us for select line to go fully low after selecting chip . 
     
     return1 = dac_port.write(data >> 8);// We have a 8 bit interface but need to write 16 bits. Two write in sequence (or use "transfer" instead)
     return2 = dac_port.write(data);
     
-    wait(0.0001); // wait 100us for select line to go fully high. 
+    wait_us(1000); // wait 100us for select line to go fully high. 
     sync = 1;
 
-    wait (0.001); // changed from 0.5 , DGC, May 2019
+    wait_us(1000); // changed from 0.5 , DGC, May 2019
     return ( return2 | (return1 << 8) );
 
 }
 
 ////////////////////////////////////////////////////////
-int mySPIRead(Serial &serialOut,  unsigned short data, SPI &dac_port, DigitalOut &sync)
+int mySPIRead(unsigned short data, SPI &dac_port, DigitalOut &sync, bool scan_mode)
 // This is the basic function used to send the signals.
 {
     unsigned short nopdata= 0x0000;
     unsigned short return1, return2;
-    wait (0.05);
+    wait_us(50000);
 
     sync = 0;
     
     dac_port.write(data >> 8);// We have a 8 bit interface but need to write 16 bits. Two write in sequence (or use "transfer" instead)
     dac_port.write(data);
     
-    wait(0.000005); //
+    wait_us(5); //
  //   sync = 1;
     
  //   wait (0.00002);
@@ -88,10 +90,11 @@ int mySPIRead(Serial &serialOut,  unsigned short data, SPI &dac_port, DigitalOut
     return1= dac_port.write(nopdata >> 8);// No operation. Used to flush the 16 bits.
     return2= dac_port.write(nopdata);
     
-    wait(0.000005); //
+    wait_us(5); //
     sync = 1;
-    wait (0.05);
-    serialOut.printf("Returned data = %d (0x%04x), %d (0x%04x).\n", return1, return1, return2, return2);
+    wait_us(50000);
+    if (!scan_mode)
+        printf("Returned data = %d (0x%04x), %d (0x%04x).\n", return1, return1, return2, return2);
     return ( return2 | (return1 << 8) );    
 }
