+++ spi_functions.h
@@ -1,7 +1,7 @@
 #include "mbed.h"
 
 
-int sendConfiguration(Serial &serialOut,  SPI &dac_port, DigitalOut &sync);
+int sendConfiguration(SPI &dac_port, DigitalOut &sync, bool scan_mode);
 
 int sendTransfer(unsigned short data,  SPI &dac_port, DigitalOut &sync);
 
@@ -11,4 +11,4 @@ int powerMode(SPI &dac_port, DigitalOut &sync, int mode);
 
 int setPort(SPI &dac_port, DigitalOut &sync, int myPort, int newState);
 
-int mySPIRead(Serial &serialOut,  unsigned short data, SPI &dac_port, DigitalOut &sync);
+int mySPIRead(unsigned short data, SPI &dac_port, DigitalOut &sync, bool scan_mode);
